use retrier in ISequencer
